style: renamed to vault-docs-landing, not abstracted yet
