+++ src/views/_proxied-dot-io/vault-docs-landing/components/featured-card/style.module.css
