+++ src/views/_proxied-dot-io/vault-docs-landing/vault-content.js
