+++ src/views/_proxied-dot-io/vault-docs-landing/vault-get-started-diagram.png
